requirements.txt
streamlit
PyMuPDF
sentence-transformers
faiss-cpu
transformers
langchain
python-dotenv
openai